+++ file_shredder/lib/file_shredder/reassembler.ex
@@ -75,7 +75,7 @@ defmodule FileShredder.Reassembler do
   def reassemble(dirpath, password) do
     hashkey = Utils.Crypto.gen_key(password)
     seq_map = Path.wildcard(dirpath)
-    |> Utils.Parallel.pmap(&start_reassem(&1, hashkey))
+    |> Utils.Parallel.pooled_map(&start_reassem(&1, hashkey))
     |> Stream.filter(&valid_hmac?(&1)) # filter out invalid hmacs
     |> Enum.reduce(%{}, &gen_seq_map(&1, &2)) # reduce into sequence map
 
@@ -97,7 +97,7 @@ defmodule FileShredder.Reassembler do
     |> Enum.to_list()
     |> Stream.reject(&(n - dummy_count <= &1)) # throw out dummy fragments
     |> Stream.map(&{&1, Map.get(seq_map, gen_seq_hash(&1, hashkey))})
-    |> Utils.Parallel.pmap(&finish_reassem(&1, hashkey, file_name, chunk_size))
+    |> Utils.Parallel.pooled_map(&finish_reassem(&1, hashkey, file_name, chunk_size))
 
     Utils.File.clear_dir(dirpath)
     { :ok, file_name }
